refactor: test 클래스 용도에 맞게 분리
