+++ src/test/java/racingcar/modeltest/CarsTest.java
@@ -1,0 +1,20 @@
+package racingcar.modeltest;
+
+import static org.assertj.core.api.Assertions.*;
+
+import java.util.Arrays;
+import java.util.List;
+
+import org.junit.jupiter.api.Test;
+
+import racingcar.model.Car;
+import racingcar.model.Cars;
+
+public class CarsTest {
+	@Test
+	public void 우승자_찾기() {
+		Cars cars = new Cars(Arrays.asList(new Car("배카라", 4), new Car("아스피", 5), new Car("아놀드", 3)));
+		List<String> result = cars.findWinnerCars();
+		assertThat(result).isEqualTo(Arrays.asList("아스피"));
+	}
+}
